requirements.txt
django
django-filter
djangorestframework
djangorestframework-jwt
markdown
psycopg2
